+++ contrib/react-sticker/sticker.jsx
@@ -1,7 +1,7 @@
 import { NodeView } from '@bangle.dev/core';
 import { Node } from '@bangle.dev/core/prosemirror/model';
 import { domSerializationHelpers } from '@bangle.dev/core/utils/dom-serialization-helpers';
-import { keymap } from '@bangle.dev/core/utils/keymap';
+import { keymap } from 'prosemirror-keymap';
 import PropTypes from 'prop-types';
 import React from 'react';
 import brontosaurusImg from './img/brontosaurus.png';