+++ contrib/text-formatting/superscript.js
@@ -1,6 +1,6 @@
 import { toggleMark } from '@bangle.dev/core/prosemirror/commands';
 import { isMarkActiveInSelection } from '@bangle.dev/core/utils/pm-utils';
-import { keymap } from '@bangle.dev/core/utils/keymap';
+import { keymap } from 'prosemirror-keymap';
 
 export const spec = specFactory;
 export const plugins = pluginsFactory;